pyproject.toml
[tool.poetry]
name = "SpaceX_DS_ML"
version = "0.1.0"
description = ""
authors = ["Yueren Xing <yxing922@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
symspellpy = "~=6.7.8"
hydra-core = "~=1.3.2"
pydantic = "~=2.8.2"
mlflow = "~=2.16.0"
dvc = {version = "~=3.55.2", extras=["gdrive", "gs"]}
joblib = "~=1.4.2"
tensorflow = "~=2.17.0"
scikit-learn = "~=1.5.1"
pandas = "~=2.2.2"
dask = {version = "~=2024.8.2", extras=["complete"]}
numpy = "~=1.26.0"
scipy = "~=1.14.1"
statsmodels = "~=0.14.2"
arch = "~=7.0.0"
prophet = "~=1.1.5"
nltk = "~=3.9.1"
xgboost = "2.1.1"
catboost = {extras = ["gpu"], version = "^1.2.5"}
lightgbm = "~=4.5.0"
matplotlib = "~=3.9.2"
seaborn = "~=0.13.2"
plotly = "~=5.24.0"
tqdm = "~=4.66.5"
missingno = "~=0.5.2"
feature-engine = "~=1.8.1"
optuna = "~=4.0.0"
shap = "~=0.46.0"
eli5 = "~=0.13.0"
transformers = {version = "~=4.44.2", extras=["tensorflow"]}
tokenizers = "~=0.19.0"
fsspec = {version = "~=2024.9.0", extras=["gcs"]}
gcsfs = "~=2024.9"
google-cloud-secret-manager = "~=2.20.2"
fastparquet = "~=2024.5"
distributed = "~=2024.8"
dask-cloudprovider = {version = "2022.10", extras=["gcp"]}
dask-ml = "~=2024.4"

[tool.poetry.group.dev.dependencies]
psutil = "~=6.0.0"
pytest = "~=8.3.2"
black = "~=24.8.0"
isort = "~=5.13.2"
flake8 = "~=7.1.1"
mypy = "~=1.11.2"
jupyterlab = "~=4.2.5"
types-PyYAML = "~=6.0.12.20240808"
types-cachetools = "~=5.5.0.20240820"
types-decorator = "~=5.1.8.20240310"
types-protobuf = "~=5.27.0.20240907"
types-pycurl = "~=7.45.3.20240421"
types-requests = "~=2.32.0.20240907"
types-six = "~=1.16.21.20240513"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
indent = 4
lines_between_types = 1
known_first_party = ["src"]
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.12"
show_error_codes = true
ignore_missing_imports = true
install_types = true
non_interactive = true
follow_imports = "silent"
no_strict_optional = true

warn_redundant_casts = false
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_no_return = true
warn_incomplete_stub = true

disallow_subclassing_any = true

disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = true

verbosity = 0
